requirements.txt
aws-cdk-lib==2.72.1
constructs>=10.0.0,<11.0.0
#cloud-assembly-schema==2.72.1
